requirements.txt
pyaudio
openai 
numpy
setuptools
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client